refactor(chat): extract pure utility functions
Extract three pure utility functions from chat.ts into separate modules:
- mightBeCompleteJson: JSON bracket balance heuristic
- messagesContainImages: vision content detection
- countInputImages: image counting for cost calculation

This is phase 1 of the chat.ts modularization effort.
